fix(ui): split header into responsive components to prevent overflow
