+++ apps/frontend/src/view/pages/assets.page.tsx
@@ -25,7 +25,8 @@ import { IconButton } from "../components/icon-button.component";
 import { Page } from "../components/page.component";
 import { SearchField } from "../components/search-field.component";
 import { CustomTableHeaderCell } from "../components/table-header.component";
-import { CreatePage, HeaderNavigation } from "../components/with-menu.component";
+import { CreatePage } from "../components/create-page.component";
+import { HeaderUtilityControls } from "../components/header-utility-controls.component";
 import { withProject } from "../components/with-project.hoc";
 import AssetDialogPage from "./asset-dialog.page";
 import { useConfirm } from "../../application/hooks/use-confirm.hook";
@@ -510,4 +511,4 @@ const AssetTableRow = ({ asset, onEdit, onDelete, language: _language, userRole 
     );
 };
 
-export const AssetsPage = CreatePage(HeaderNavigation, withProject(AssetsPageBody), true);
+export const AssetsPage = CreatePage(HeaderUtilityControls, withProject(AssetsPageBody), true);
