+++ apps/frontend/src/view/pages/privacy-policy.page.tsx
@@ -3,7 +3,8 @@ import { useLayoutEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { NavigationActions } from "../../application/actions/navigation.actions";
 import { Page } from "../components/page.component";
-import { CreatePage, HeaderNavigation } from "../components/with-menu.component";
+import { CreatePage } from "../components/create-page.component";
+import { HeaderUtilityControls } from "../components/header-utility-controls.component";
 import { useGetMarkdownText } from "../wrappers/markdown.wrapper";
 import Markdown from "markdown-to-jsx";
 import { useAppDispatch } from "#application/hooks/use-app-redux.hook.ts";
@@ -56,4 +57,4 @@ const PrivacyPolicyPageBody = () => {
     );
 };
 
-export const PrivacyPolicyPage = CreatePage(HeaderNavigation, PrivacyPolicyPageBody);
+export const PrivacyPolicyPage = CreatePage(HeaderUtilityControls, PrivacyPolicyPageBody);
