+++ apps/frontend/src/view/pages/threats.page.tsx
@@ -19,7 +19,8 @@ import { IconButton } from "../components/icon-button.component";
 import { Page } from "../components/page.component";
 import { SearchField } from "../components/search-field.component";
 import { CustomTableHeaderCell } from "../components/table-header.component";
-import { CreatePage, HeaderNavigation } from "../components/with-menu.component";
+import { CreatePage } from "../components/create-page.component";
+import { HeaderUtilityControls } from "../components/header-utility-controls.component";
 import ThreatDialogPage from "./threat-dialog.page";
 import { withProject } from "../components/with-project.hoc";
 import { useAppDispatch, useAppSelector } from "#application/hooks/use-app-redux.hook.ts";
@@ -550,4 +551,4 @@ const CustomTableCell = ({ sx, showBorder = false, children, ...props }: CustomT
     );
 };
 
-export const ThreatsPage = CreatePage(HeaderNavigation, withProject(ThreatsPageBody), true);
+export const ThreatsPage = CreatePage(HeaderUtilityControls, withProject(ThreatsPageBody), true);
